requirements.txt
browserstack-local >= 1.2.3
selenium
browserstack-sdk
google-cloud-translate
python-dotenv